Move useGrid to its own file
